Output: The locked stub code in your editor will print the integer returned by your getHeight function denoting the height of the BST
The height of a binary search tree is the number of edges between the tree's root and its furthest leaf. You are given a pointer, "root", pointing to the root of a binary search tree. Complete the getHeight function provided in your editor so that it returns the height of the binary search tree.