Create Day2_Compound_Event_Probability.txt
In this challenge, we practice calculating the probability of a compound event. We recommend you review today's Probability Tutorial before attempting this challenge.
There are 3  urns labeled x, y,and z.


Urn  X contains  4 red balls and 3 black balls.
Urn Y  contains  5 red balls and 4 black balls.
Urn Z  contains  4 red balls and 4 black balls.

One ball is drawn from each of the  urns. What is the probability that, of the  balls drawn,  are red and  is black?