Create TheCentralLimitTheoremI.java
Input Format

There are "4"  lines of input (shown below):

9800
49
205
15
The first line contains the maximum weight the elevator can transport. The second line contains the number of boxes in the cargo. The third line contains the mean weight of a cargo box, and the fourth line contains its standard deviation.

If you do not wish to read this information from stdin, you can hard-code it into your program.