Create TheCentralLimitTheoremll.java
There are "4"  lines of input (shown below):

250
100
2.4
2.0
The first line contains the number of last-minute tickets available at the box office. The second line contains the number of students waiting to buy tickets. The third line contains the mean number of purchased tickets, and the fourth line contains the standard deviation.

If you do not wish to read this information from stdin, you can hard-code it into your program.