Create TheCentralLimitTheorem_III.java
There are five lines of input (shown below):

100
500
80
.95
1.96
The first line contains the sample size. The second and third lines contain the respective mean () and standard deviation (). The fourth line contains the distribution percentage we want to cover (as a decimal), and the fifth line contains the value of .

If you do not wish to read this information from stdin, you can hard-code it into your program.